pyproject.toml
[project]
name = "yt-transcribe"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "openai-whisper>=20231117",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
